requirements.txt
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
apturl==0.5.2
asttokens==2.4.1
async-timeout==4.0.2
attrs==22.2.0
avidtools==0.1.1.2
backcall==0.2.0
backoff==2.2.1
base2048==0.1.3
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.1.0
blinker==1.7.0
blobfile==2.0.1
boto3==1.34.63
botocore==1.34.63
Brlapi==0.8.3
cachetools==5.3.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.0.1
chatgpt==2.2212.0
chevron==0.14.0
click==8.1.7
click-default-group==1.2.4
cmd2==2.4.3
cohere==4.56
colorama==0.4.6
comm==0.2.2
command-not-found==0.3
commonmark==0.9.1
cryptography==42.0.5
cupshelpers==1.0
dacite==1.6.0
dataclasses-json==0.6.4
datasets==2.16.1
DateTime==5.4
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
deepl==1.17.0
defer==1.0.6
defusedxml==0.7.1
dill==0.3.7
discord==2.3.2
discord.py==2.3.2
distro==1.9.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
docopt==0.6.2
ecoji==0.1.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
fastavro==1.9.4
fastjsonschema==2.19.1
filelock==3.9.0
Flask==3.0.2
frozenlist==1.3.3
fschat==0.2.36
fsspec==2023.10.0
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
greenlet==2.0.1
groq==0.5.0
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.4
httplib2==0.20.2
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.3
importlib-metadata==6.11.0
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
ipykernel==6.29.3
ipython==8.12.3
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-path==0.3.2
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
keyring==23.5.0
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-text-splitters==0.0.1
langsmith==0.1.25
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.10.0
litellm==1.35.5
llm==0.13.1
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Markdown==3.6
markdown2==2.4.13
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.6
mistune==3.0.2
more-itertools==8.10.0
motor==3.3.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nemollm==0.3.5
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.2.1
nh3==0.2.15
nltk==3.8.1
numpy==1.26.4
nvdlib==0.7.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
octoai-sdk==0.10.1
olefile==0.46
openai==1.12.0
openapi==1.1.0
openapi-core==0.19.0
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
openpyxl==3.1.2
orjson==3.9.15
packaging==23.2
pandas==2.2.1
pandas-stubs==2.2.0.240218
pandocfilters==1.5.1
parse==1.20.1
parso==0.8.3
pathable==0.4.3
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
pip-review==1.3.0
pipreqs==0.5.0
platformdirs==4.2.0
playwright==1.28.0
pluggy==1.4.0
prance==23.6.21.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybars4==0.9.13
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.17
pycups==2.0.1
pydantic==2.6.1
pydantic_core==2.16.2
pyee==9.0.4
Pygments==2.13.0
PyGObject==3.42.1
PyInquirer==1.0.3
PyJWT==2.3.0
pymacaroons==0.13.0
PyMeta3==0.5.1
pymongo==4.6.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
pyRFC3339==1.1
pyrsistent==0.19.3
pytest==8.1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-multipart==0.0.9
python-ulid==2.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0
pyzmq==25.1.2
rapidfuzz==3.6.2
referencing==0.31.1
regex==2023.12.25
replicate==0.24.0
reportlab==3.6.8
requests==2.31.0
requests-futures==1.0.1
requests-mock==1.12.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==12.6.0
rpds-py==0.18.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
safetensors==0.4.2
scipy==1.12.0
SecretStorage==3.3.1
semantic-kernel==0.9.4b1
sentencepiece==0.2.0
shortuuid==1.0.13
six==1.16.0
slack_sdk==3.27.1
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
SQLAlchemy==2.0.28
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.36
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.36.3
stdlibs==2024.1.28
svgwrite==1.4.3
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.6.0
tinycss2==1.2.1
tls-client==1.0.1
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.1
tornado==6.4
tqdm==4.64.1
traitlets==5.14.1
transformers==4.38.2
triton==2.2.0
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.20240311
types-requests==2.31.0.20240311
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.28.0
wadllib==1.3.6
wavedrom==2.0.3.post3
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==3.0.1
xdg==5
xkit==0.0.0
xxhash==3.4.1
yarg==0.1.9
yarl==1.8.2
zalgolib==0.2.2
zipp==1.0.0
zope.interface==6.2
